requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
hubspot-api-client==11.1.0
python-dotenv==1.0.1
